+++ openeo/rest/vectorcube.py
@@ -6,7 +6,7 @@ from openeo.internal.documentation import openeo_process
 from openeo.internal.graph_building import PGNode
 from openeo.metadata import CollectionMetadata
 from openeo.rest._datacube import _ProcessGraphAbstraction
-from openeo.rest.job import RESTJob
+from openeo.rest.job import BatchJob
 from openeo.util import legacy_alias, dict_no_none
 
 if typing.TYPE_CHECKING:
@@ -77,7 +77,7 @@ class VectorCube(_ProcessGraphAbstraction):
             self,
             outputfile: Union[str, pathlib.Path], out_format: str = None,
             print=print, max_poll_interval=60, connection_retry_interval=30,
-            job_options=None, **format_options) -> RESTJob:
+            job_options=None, **format_options) -> BatchJob:
         """
         Evaluate the process graph by creating a batch job, and retrieving the results when it is finished.
         This method is mostly recommended if the batch job is expected to run in a reasonable amount of time.
@@ -97,7 +97,7 @@ class VectorCube(_ProcessGraphAbstraction):
             print=print, max_poll_interval=max_poll_interval, connection_retry_interval=connection_retry_interval
         )
 
-    def create_job(self, out_format=None, job_options=None, **format_options) -> RESTJob:
+    def create_job(self, out_format=None, job_options=None, **format_options) -> BatchJob:
         """
         Sends a job to the backend and returns a ClientJob instance.