+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Rename `RESTJob` class name to less cryptic and more user-friendly `BatchJob`. 
+  Original `RESTJob` is still available as deprecated alias.
+  ([#280](https://github.com/Open-EO/openeo-python-client/issues/280))
+
+
 ### Removed
 
 ### Fixed